+++ src/app/[locale]/metrics/_components/Analytics/GBFSFeedAnalytics/DetailPanel.tsx
