+++ src/app/[locale]/metrics/_components/Analytics/GBFSFeedAnalytics/GBFSFeedAnalyticsTable.tsx
