+++ src/app/[locale]/metrics/_components/Analytics/GTFSFeatureAnalytics/GTFSFeatureAnalytics.tsx
@@ -1,57 +1,83 @@
+'use client';
+
 import { useState, useEffect, useMemo } from 'react';
 import { useSearchParams, useRouter } from 'next/navigation';
-
 import {
   MaterialReactTable,
-  useMaterialReactTable,
   type MRT_ColumnDef,
+  type MRT_Cell,
+  useMaterialReactTable,
 } from 'material-react-table';
 import {
   XAxis,
   YAxis,
   CartesianGrid,
+  Tooltip,
   Legend,
   ResponsiveContainer,
   Brush,
   Line,
   LineChart,
-  Tooltip,
 } from 'recharts';
 import Box from '@mui/material/Box';
-
-import { Typography, Button, Alert, AlertTitle } from '@mui/material';
+import {
+  Typography,
+  Button,
+  IconButton,
+  Alert,
+  AlertTitle,
+} from '@mui/material';
 import * as React from 'react';
 import { useTheme } from '@mui/material/styles';
 import { InfoOutlined, ListAltOutlined } from '@mui/icons-material';
-import { type GBFSNoticeMetrics } from '../types';
-import { useRemoteConfig } from '../../../context/RemoteConfigProvider';
+import { type FeatureMetrics } from '../types';
+import { useRemoteConfig } from '../../../../../context/RemoteConfigProvider';
+import MUITooltip from '@mui/material/Tooltip';
+import { GTFS_ORG_LINK } from '../../../../../constants/Navigation';
+import {
+  DATASET_FEATURES,
+  getComponentDecorators,
+} from '../../../../../utils/consts';
 
-export default function GBFSNoticeAnalytics(): React.ReactElement {
+export default function GTFSFeatureAnalytics(): React.ReactElement {
   const router = useRouter();
   const searchParams = useSearchParams();
-  const noticeCode = searchParams.get('noticeCode');
-  const [data, setData] = useState<GBFSNoticeMetrics[]>([]);
+  const featureName = searchParams.get('featureName');
+  const [data, setData] = useState<FeatureMetrics[]>([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
   const { config } = useRemoteConfig();
+
+  const getUniqueKeyStringValues = (key: keyof FeatureMetrics): string[] => {
+    const subGroups = new Set<string>();
+    data.forEach((item) => {
+      if (item[key] !== undefined) {
+        subGroups.add(item[key] as string);
+      }
+    });
+    return Array.from(subGroups);
+  };
 
   useEffect(() => {
     const fetchData = async (): Promise<void> => {
       try {
         const response = await fetch(
-          `${config.gbfsMetricsBucketEndpoint}/notices_metrics.json`,
+          `${config.gtfsMetricsBucketEndpoint}/features_metrics.json`,
         );
         if (!response.ok) {
           throw new Error('Network response was not ok');
         }
         const fetchedData = await response.json();
-        const dataWLatestCount = fetchedData.map(
-          (notice: GBFSNoticeMetrics) => ({
-            ...notice,
-            latest_feed_count: notice.feeds_count.slice(-1)[0],
-          }),
-        );
-        setData(dataWLatestCount);
+        const dataWithGroups = fetchedData.map((feature: FeatureMetrics) => {
+          return {
+            ...feature,
+            latest_feed_count: feature.feeds_count.slice(-1)[0],
+            feature_group: DATASET_FEATURES[feature.feature]?.component,
+            feature_sub_group:
+              DATASET_FEATURES[feature.feature]?.componentSubgroup,
+          };
+        });
+        setData(dataWithGroups);
       } catch (error) {
         if (error instanceof Error) {
           setError(error.message);
@@ -65,21 +91,85 @@ export default function GBFSNoticeAnalytics(): React.ReactElement {
     void fetchData();
   }, []);
 
-  const columns = useMemo<Array<MRT_ColumnDef<GBFSNoticeMetrics>>>(
+  const columns = useMemo<Array<MRT_ColumnDef<FeatureMetrics>>>(
     () => [
       {
-        accessorKey: 'keyword',
-        header: 'Keyword',
+        accessorKey: 'feature',
+        header: 'Feature Name',
         size: 300,
+        enableClickToCopy: true,
+        Cell: ({
+          cell,
+          renderedCellValue,
+        }: {
+          cell: MRT_Cell<FeatureMetrics>;
+          renderedCellValue: React.ReactNode;
+        }) => {
+          return (
+            <div>
+              {renderedCellValue}
+              <MUITooltip
+                title={`View ${cell.getValue<string>()} definition`}
+                arrow
+              >
+                <IconButton
+                  onClick={() => {
+                    window.open(
+                      `${GTFS_ORG_LINK}/getting_started/features/base_add-ons/#${cell
+                        .getValue<string>()
+                        .toLowerCase()}`,
+                      '_blank',
+                    );
+                  }}
+                >
+                  <InfoOutlined />
+                </IconButton>
+              </MUITooltip>
+            </div>
+          );
+        },
       },
       {
-        accessorKey: 'gbfs_file',
-        header: 'GBFS File',
-        size: 150,
+        accessorKey: 'feature_group',
+        header: 'Feature Group',
+        size: 200,
+        filterVariant: 'multi-select',
+        filterSelectOptions: getUniqueKeyStringValues('feature_group'),
+        Cell: ({ cell }: { cell: MRT_Cell<FeatureMetrics> }) => {
+          const group = cell.getValue<string>();
+          return group == null ? null : (
+            <span
+              style={{
+                backgroundColor: getComponentDecorators(group).color,
+                borderRadius: '5px',
+                padding: '2px 8px',
+              }}
+            >
+              {group}
+            </span>
+          );
+        },
       },
       {
-        accessorKey: 'schema_path',
-        header: 'Schema Path',
+        accessorKey: 'feature_sub_group',
+        header: 'Feature Sub Group',
+        size: 200,
+        filterVariant: 'multi-select',
+        filterSelectOptions: getUniqueKeyStringValues('feature_sub_group'),
+        Cell: ({ cell }: { cell: MRT_Cell<FeatureMetrics> }) => {
+          const group = cell.getValue<string>();
+          return group == null ? null : (
+            <span
+              style={{
+                backgroundColor: getComponentDecorators(group).color,
+                borderRadius: '5px',
+                padding: '2px 8px',
+              }}
+            >
+              {group}
+            </span>
+          );
+        },
       },
       {
         accessorKey: 'latest_feed_count',
@@ -89,7 +179,7 @@ export default function GBFSNoticeAnalytics(): React.ReactElement {
         muiFilterSliderProps: {
           marks: true,
           max: data.reduce(
-            (max, notice) => Math.max(max, notice.feeds_count.slice(-1)[0]),
+            (max, feature) => Math.max(max, feature.latest_feed_count),
             0,
           ),
           min: 0,
@@ -101,11 +191,11 @@ export default function GBFSNoticeAnalytics(): React.ReactElement {
   );
 
   const initialFilters =
-    noticeCode != null
+    featureName != null
       ? [
           {
-            id: 'keyword',
-            value: noticeCode,
+            id: 'feature',
+            value: featureName,
           },
         ]
       : [];
@@ -115,9 +205,9 @@ export default function GBFSNoticeAnalytics(): React.ReactElement {
     data,
     initialState: {
       showColumnFilters: true,
-      columnPinning: { left: ['mrt-row-expand', 'keyword'] },
+      columnPinning: { left: ['mrt-row-expand', 'feature'] },
       density: 'compact',
-      sorting: [{ id: 'keyword', desc: false }],
+      sorting: [{ id: 'feature', desc: false }],
       columnFilters: initialFilters,
       expanded: initialFilters.length > 0 ? true : {},
     },
@@ -135,7 +225,9 @@ export default function GBFSNoticeAnalytics(): React.ReactElement {
       const metrics = row.original;
 
       const chartData = metrics.computed_on.map((date, index) => ({
-        date: new Date(date).toLocaleDateString('en-CA', { timeZone: 'UTC' }),
+        date: new Date(date).toLocaleDateString('en-CA', {
+          timeZone: 'UTC',
+        }),
         feeds: metrics.feeds_count[index],
       }));
       const domain = [
@@ -152,7 +244,7 @@ export default function GBFSNoticeAnalytics(): React.ReactElement {
           }}
         >
           <Box sx={{ flex: 1, paddingRight: 2 }}>
-            <Typography gutterBottom>Monthly Notice Metrics</Typography>
+            <Typography gutterBottom>Monthly Feature Metrics</Typography>
             <ResponsiveContainer width='100%' height={300}>
               <LineChart data={chartData}>
                 <CartesianGrid strokeDasharray='3 3' />
@@ -181,8 +273,8 @@ export default function GBFSNoticeAnalytics(): React.ReactElement {
                     verticalAlign: 'middle',
                   }}
                 />
-                This graph shows the monthly feed validation metrics, including
-                the count of feeds associated with each notice over time.
+                This graph shows the monthly feed metrics, including the count
+                of feeds associated with each feature over time.
               </Typography>
               <Button
                 variant='contained'
@@ -191,9 +283,7 @@ export default function GBFSNoticeAnalytics(): React.ReactElement {
                 startIcon={<ListAltOutlined />}
                 onClick={() => {
                   router.push(
-                    `/metrics/gbfs/feeds?schemaPath=${encodeURIComponent(
-                      metrics.schema_path,
-                    )}`,
+                    `/metrics/gtfs/feeds?featureName=${metrics.feature}`,
                   );
                 }}
               >
@@ -206,10 +296,6 @@ export default function GBFSNoticeAnalytics(): React.ReactElement {
     },
   });
 
-  if (loading) {
-    return <div>Loading...</div>;
-  }
-
   return (
     <Box sx={{ mx: 6 }}>
       <Typography
@@ -218,7 +304,7 @@ export default function GBFSNoticeAnalytics(): React.ReactElement {
         color='primary'
         sx={{ fontWeight: 700, mb: 2 }}
       >
-        GBFS Notices Metrics
+        GTFS Features Metrics{' '}
       </Typography>
       {error != null && (
         <Alert severity='error'>
