+++ src/app/[locale]/metrics/_components/Analytics/types.ts
@@ -1,0 +1,1 @@
+export * from '../../../../utils/analytics-types';
