+++ src/app/[locale]/metrics/_components/Analytics/utils.ts
