+++ src/app/[locale]/metrics/gbfs/feeds/page.tsx
@@ -1,0 +1,6 @@
+import { type ReactElement } from 'react';
+import GBFSFeedAnalytics from '../../_components/Analytics/GBFSFeedAnalytics/GBFSFeedAnalytics';
+
+export default function GBFSFeedsMetricsPage(): ReactElement {
+  return <GBFSFeedAnalytics />;
+}
