+++ src/app/[locale]/metrics/gbfs/notices/page.tsx
@@ -1,0 +1,6 @@
+import { type ReactElement } from 'react';
+import GBFSNoticeAnalytics from '../../_components/Analytics/GBFSNoticeAnalytics/GBFSNoticeAnalytics';
+
+export default function GBFSNoticesMetricsPage(): ReactElement {
+  return <GBFSNoticeAnalytics />;
+}
