+++ src/app/[locale]/metrics/gbfs/versions/page.tsx
@@ -1,0 +1,6 @@
+import { type ReactElement } from 'react';
+import GBFSVersionAnalytics from '../../_components/Analytics/GBFSVersionAnalytics/GBFSVersionAnalytics';
+
+export default function GBFSVersionsMetricsPage(): ReactElement {
+  return <GBFSVersionAnalytics />;
+}
