+++ src/app/[locale]/metrics/gtfs/features/page.tsx
@@ -1,0 +1,6 @@
+import { type ReactElement } from 'react';
+import GTFSFeatureAnalytics from '../../_components/Analytics/GTFSFeatureAnalytics/GTFSFeatureAnalytics';
+
+export default function GTFSFeaturesMetricsPage(): ReactElement {
+  return <GTFSFeatureAnalytics />;
+}
