+++ src/app/[locale]/metrics/gtfs/feeds/page.tsx
@@ -1,0 +1,6 @@
+import { type ReactElement } from 'react';
+import GTFSFeedAnalytics from '../../_components/Analytics/GTFSFeedAnalytics/GTFSFeedAnalytics';
+
+export default function GTFSFeedsMetricsPage(): ReactElement {
+  return <GTFSFeedAnalytics />;
+}
