+++ src/app/[locale]/metrics/gtfs/notices/page.tsx
@@ -1,0 +1,6 @@
+import { type ReactElement } from 'react';
+import GTFSNoticeAnalytics from '../../_components/Analytics/GTFSNoticeAnalytics/GTFSNoticeAnalytics';
+
+export default function GTFSNoticesMetricsPage(): ReactElement {
+  return <GTFSNoticeAnalytics />;
+}
