+++ src/app/[locale]/metrics/gtfs/page.tsx
@@ -1,0 +1,6 @@
+import { type ReactElement } from 'react';
+import GTFSFeedAnalytics from '../_components/Analytics/GTFSFeedAnalytics/GTFSFeedAnalytics';
+
+export default function GTFSMetricsPage(): ReactElement {
+  return <GTFSFeedAnalytics />;
+}
