+++ src/app/[locale]/metrics/layout.tsx
@@ -1,0 +1,10 @@
+import { type ReactElement } from 'react';
+import { ReduxGateWrapper } from '../../components/ReduxGateWrapper';
+
+export default function MetricsLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}): ReactElement {
+  return <ReduxGateWrapper>{children}</ReduxGateWrapper>;
+}
