+++ src/app/store/saga/gbfs-analytics-saga.ts
@@ -11,9 +11,9 @@ import {
   type AnalyticsFile,
   type GBFSFeedMetrics,
   type GBFSMetrics,
-} from '../../screens/Analytics/types';
+} from '../../utils/analytics-types';
 import { getLocationName } from '../../services/feeds/utils';
-import { getAnalyticsBucketEndpoint } from '../../screens/Analytics/utils';
+import { getAnalyticsBucketEndpoint } from '../../[locale]/metrics/_components/Analytics/utils';
 
 function* fetchFeedMetricsSaga(
   action: ReturnType<typeof selectFile>,
